Move worker cache service to node core
