+++ packages/node-core/src/indexer/worker/worker.cache.service.ts
@@ -1,6 +1,8 @@
 // Copyright 2020-2023 SubQuery Pte Ltd authors & contributors
 // SPDX-License-Identifier: GPL-3.0
 
+import {isMainThread} from 'worker_threads';
+import {Injectable} from '@nestjs/common';
 import {Cache} from '@subql/types-core';
 import {unwrapProxyArgs} from './utils';
 
@@ -23,4 +25,17 @@ export function cacheHostFunctions(cache: Cache): HostCache {
     cacheGet: cache.get.bind(cache),
     cacheSet: cache.set.bind(cache),
   };
+}
+
+@Injectable()
+export class WorkerInMemoryCacheService {
+  constructor(private host: HostCache) {
+    if (isMainThread) {
+      throw new Error('Expected to be worker thread');
+    }
+  }
+
+  getCache(): Cache {
+    return hostCacheToCache(this.host);
+  }
 }
